app updated with scopes changes
