+++ api-keys-auth-sample/README.md
@@ -7,7 +7,7 @@
 
 ### Run your server & client app on same port
 
-1. `cd client-to-server-sample/`
+1. `cd api-keys-auth-sample/`
 2. edit `[YOUR_API_KEY]` with your secret key in `server.js`
 3. edit `[YOUR_CLIENT_ID]` with your client id in `client_application/app.js`
 3. run `npm install`
