+++ api-keys-auth-sample/client_application/index.html
