+++ api-keys-auth-sample/client_application/service.js
