+++ api-keys-auth-sample/package.json
@@ -6,7 +6,6 @@
   "repository": "git@github.com/Livestream/livestream-api-samples",
   "devDependencies": {
     "crypto": "^0.0.3",
-    "request": "^2.74.0",
     "express": "^4.14.0",
     "moment": "^2.14.1"
   }
