+++ secure-token-auth-sample/client_application/accounts.html
