+++ secure-token-auth-sample/client_application/app.js
