+++ secure-token-auth-sample/client_application/index.html
