+++ secure-token-auth-sample/client_application/service.js
