+++ secure-token-auth-sample/package.json
